requirements.txt
hikyuu
tushare
akshare
pyyaml
loguru
pytest
pandas
numpy
ta-lib
python-dotenv
click
